requirements.txt
pandas==3.0.2
numpy==2.2.4
scikit-learn==1.8.0
xgboost==3.2.0
imbalanced-learn==0.14.1
pyyaml==6.0.3
joblib==1.5.3
matplotlib==3.10.9
seaborn==0.13.2
